This all data that I can find.
About alimentation.  As I understand it, this dog food is a must have. I can't find accurate information on each breed. I change it to how this dog treats a human/
In this data, all the parameters are from several databases/You can select any one and I will change the data about the breed/